+++ cs/src/io/unsafe/StreamCloning.cs
@@ -75,57 +75,11 @@ namespace Bond.IO.Unsafe
             public const uint FILE_GENERIC_READ = 0x00120089;
 
             [DllImport("kernel32.dll", SetLastError = true)]
-#if !(NETSTANDARD1_3 || NETSTANDARD1_6)
-            [SuppressUnmanagedCodeSecurity]
-#endif
             public static extern SafeFileHandle ReOpenFile(SafeFileHandle hFile, uint access, FileShare mode, uint flags);
         }
 
         static class MemoryStreamCloner
         {
-#if !(NET46 || NETSTANDARD1_3 || NETSTANDARD1_6)
-            delegate void GetOriginAndLength(MemoryStream stream, out int origin, out int length);
-            static readonly GetOriginAndLength getOriginAndLength;
-
-            static MemoryStreamCloner()
-            {
-                // MemoryStream can be created from byte[] at non-zero origin offset which
-                // is not exposed via public APIs.
-                // TODO: handle the case when the method API is not available (e.g. copy buffer)
-                var internalGetOriginAndLength = typeof(MemoryStream).GetMethod(
-                    "InternalGetOriginAndLength", BindingFlags.NonPublic | BindingFlags.Instance);
-
-                var stream = Expression.Parameter(typeof(MemoryStream));
-                var origin = Expression.Parameter(typeof(int).MakeByRefType());
-                var length = Expression.Parameter(typeof(int).MakeByRefType());
-
-                //internal method InternalGetOriginAndLength exists in .NET
-                if (internalGetOriginAndLength != null)
-                {
-                    getOriginAndLength = Expression.Lambda<GetOriginAndLength>(
-                        Expression.Call(stream, internalGetOriginAndLength, origin, length), stream, origin, length)
-                        .Compile();
-                }
-                else
-                {
-                    //Mono uses 'initialIndex' and 'length' fields in memory stream to handle the data
-                    //see ref: https://github.com/mono/mono/blob/989301647e7cdf93acb69aacabcbb2bb2884e041/mcs/class/corlib/System.IO/MemoryStream.cs#L122-L126
-                    var assign1 = Expression.Assign(origin, Expression.Field(stream, "initialIndex")); //origin = stream.initialIndex
-                    var assign2 = Expression.Assign(length, Expression.Field(stream, "length")); //length = stream.length
-
-                    getOriginAndLength = Expression.Lambda<GetOriginAndLength>(
-                        Expression.Block(assign1, assign2), stream, origin, length)
-                        .Compile();
-                }
-            }
-
-            internal static MemoryStream CloneMemoryStream(MemoryStream stream)
-            {
-                int origin, length;
-                getOriginAndLength(stream, out origin, out length);
-                return new MemoryStream(stream.GetBuffer(), origin, length - origin, false, true) { Position = stream.Position };
-            }
-#else // NET46 & NETSTANDARD implementation
             internal static MemoryStream CloneMemoryStream(MemoryStream stream)
             {
                 ArraySegment<byte> buffer;
@@ -137,7 +91,6 @@ namespace Bond.IO.Unsafe
 
                 return new MemoryStream(buffer.Array, buffer.Offset, buffer.Count, false, true) { Position = stream.Position };
             }
-#endif
         }
     }
 }