+++ tools/ci-scripts/linux/build_cs.zsh
@@ -19,8 +19,6 @@ msbuild /p:Configuration=Debug /m $BOND_ROOT/cs/cs.sln
 msbuild /p:Configuration=Fields /m $BOND_ROOT/cs/cs.sln
 
 mono /root/NUnit.ConsoleRunner/tools/nunit3-console.exe -framework=mono -labels=All \
-    $BOND_ROOT/cs/test/core/bin/debug/Properties/net45/Bond.UnitTest.dll \
-    $BOND_ROOT/cs/test/core/bin/debug/Fields/net45/Bond.UnitTest.dll \
-    $BOND_ROOT/cs/test/coreNS10/bin/debug/Properties/net45/Bond.UnitTestCoreNS10.dll \
-    $BOND_ROOT/cs/test/coreNS10/bin/debug/Fields/net45/Bond.UnitTestCoreNS10.dll \
-    $BOND_ROOT/cs/test/internal/bin/debug/net45/Bond.InternalTest.dll
+    $BOND_ROOT/cs/test/core/bin/debug/Properties/netstandard2.0/Bond.UnitTest.dll \
+    $BOND_ROOT/cs/test/core/bin/debug/Fields/netstandard2.0/Bond.UnitTest.dll \
+    $BOND_ROOT/cs/test/internal/bin/debug/netstandard2.0/Bond.InternalTest.dll